chore(nextjs): Rename middleware folder to ssr
