--- packages/nextjs/src/middleware/index.ts
@@ -1,1 +1,0 @@
-export * from './withServerSideAuth';
