+++ packages/nextjs/src/ssr/index.ts
@@ -1,1 +1,1 @@
-export * from '../middleware';
+export * from './withServerSideAuth';
