+++ packages/nextjs/src/ssr/types.ts
