+++ packages/nextjs/src/ssr/utils/getAuthData.ts
