+++ packages/nextjs/src/ssr/utils/injectAuthIntoRequest.ts
