+++ packages/nextjs/src/ssr/utils/sanitizeAuthData.ts
