+++ packages/nextjs/src/ssr/utils/serializeProps.ts
