+++ packages/nextjs/src/ssr/withServerSideAuth.ts
